refactored appnav to have separate components in same file
