--- components/util/AuthLink.tsx
@@ -1,20 +1,0 @@
-import * as React from 'react'
-
-const AuthLink = () => (
-  <div className="navbar-nav collapse navbar-collapse">
-    <a href="#" className="nav-item nav-link active">
-      Curriculum <span className="sr-only">(current)</span>
-    </a>
-    <a href="#" className="nav-item nav-link">
-      Repo
-    </a>
-    <a href="#" className="nav-item nav-link">
-      Journey
-    </a>
-    <a href="https://chat.c0d3.com" className="nav-item nav-link">
-      Help
-    </a>
-  </div>
-)
-
-export default AuthLink
